authorization

+++ CRUD/src/routes/user.route.ts
@@ -4,6 +4,8 @@ import { validationMiddleware } from "../middleware/validator";
 import { CreateUserDTO } from "../DTO/createUser.dto";
 import { UpdateUserDTO } from "../DTO/updateUser.dto";
 import passport from "../passport";
+import { authorize } from "../middleware/auth";
+import { Role } from "../constants/role.enum";
 const router = express();
 
 /**
@@ -11,28 +13,64 @@ const router = express();
  * /user:
  *   get:
  *     summary: Returns all users
+ *     tags:
+ *       - User
+ *     security:
+ *       - bearerAuth: []
  *     responses:
  *       200:
  *         description: A successful response
+ *         content:
+ *           application/json:
+ *             schema:
+ *               $ref: '#/components/schemas/createUserResponse'
+ *       401:
+ *         description: Unauthorized
  *       404:
  *         description: Users not found
  */
+
 router.get(
   "/",
   passport.authenticate("jwt", { session: false }),
+  authorize(Role.Admin),
   userController.getAll
 );
 
+/**
+ * @openapi
+ * /user/my-detail:
+ *   get:
+ *     summary: Returns the loggedin user's detail
+ *     tags:
+ *      - User
+ *     security:
+ *       - bearerAuth: []
+ *     responses:
+ *       200:
+ *         description: A successful response
+ *         content:
+ *           application/json:
+ *             schema:
+ *               $ref: '#/components/schemas/createUserResponse'
+ *       401:
+ *         description: Unauthorized
+ */
 router.get(
-  "/myDetail",
+  "/my-detail",
   passport.authenticate("jwt", { session: false }),
+  authorize(Role.Admin, Role.User),
   userController.getMyDetail
 );
 /**
  * @openapi
  * /user/{id}:
  *   get:
- *     summary: Returns all user by Id
+ *     summary: Returns user by Id
+ *     tags:
+ *      - User
+ *     security:
+ *       - bearerAuth: []
  *     parameters:
  *       - in: path
  *         name: id
@@ -47,19 +85,24 @@ router.get(
  *               $ref: '#/components/schemas/createUserResponse'
  *       404:
  *          description: Not found
+ *       401:
+ *         description: Unauthorized
  */
 
 router.get(
   "/:id",
   passport.authenticate("jwt", { session: false }),
+  authorize(Role.Admin),
   userController.getById
 );
 
 /**
  * @openapi
- * /user:
+ * /user/signup:
  *   post:
  *     summary: Creates a new user and returns the info of the created user
+ *     tags:
+ *      - Auth
  *     requestBody:
  *       required: true
  *       content:
@@ -76,13 +119,21 @@ router.get(
  *       400:
  *         description: Bad Request Error
  */
-router.post("/", validationMiddleware(CreateUserDTO), userController.create);
+router.post(
+  "/signup",
+  validationMiddleware(CreateUserDTO),
+  userController.create
+);
 
 /**
  * @openapi
  * /user/{id}:
  *   put:
  *     summary: updates the existing user
+ *     tags:
+ *      - User
+ *     security:
+ *       - bearerAuth: []
  *     requestBody:
  *       required: true
  *       content:
@@ -98,10 +149,13 @@ router.post("/", validationMiddleware(CreateUserDTO), userController.create);
  *               $ref: '#/components/schemas/createUserResponse'
  *       400:
  *         description: Bad Request Error
+ *       401:
+ *         description: Unauthorized
  */
 router.put(
   "/:id",
   passport.authenticate("jwt", { session: false }),
+  authorize(Role.Admin),
   validationMiddleware(UpdateUserDTO),
   userController.updateById
 );
@@ -111,6 +165,8 @@ router.put(
  * /user/{id}:
  *   delete:
  *     summary: delete the user by Id
+ *     tags:
+ *      - User
  *     parameters:
  *       - in: path
  *         name: id
@@ -129,10 +185,13 @@ router.put(
  *                   example: "User of ID: c6cc6f98-cdda-45ba-a20e-6a033cb778d4 successfully deleted"
  *       404:
  *          description: User not found
+ *       401:
+ *         description: Unauthorized
  */
 router.delete(
   "/:id",
   passport.authenticate("jwt", { session: false }),
+  authorize(Role.Admin),
   userController.deleteById
 );
